Edge Calculation Update
The program can now calculate BASIC solutions to edge scrambles, but
cannot handle cycle breaks and it’s possible to put the program in a
state of infinite looping if you make an impossible piece such as a G/G
edge piece.

The ability to change the color of the centers has also been removed,
to avoid having to validate them. In the future, it’s possible to add
customization options to change the orientation for people that don’t
like white-in-front/blue-on-top.

Added a logging feature to display the moves that the calculator works
out.